refactor: rename yoopta-editor to editor
